[COMPLETE] add loading in all books

+++ Dtos/BookCategoryDto.cs
@@ -4,7 +4,7 @@ namespace BookManagement.Dtos
 {
     public class BookCategoryDto
     {
-        public int Category_Id { get; set; }
+        public long Category_Id { get; set; }
         public string Category_Name { get; set; }
         public DateTime CreatedDate { get; set; }
         public DateTime ModifiedDate { get; set; }
